fix: public test-helpers
