+++ packages/calendar/test-helpers.js
@@ -1,0 +1,3 @@
+export { CalendarObject } from './test-helpers/CalendarObject.js';
+export { DayObject } from './test-helpers/DayObject.js';
+export { weekdayNames } from './test-helpers/weekdayNames.js';
