+++ packages/calendar/test-helpers/weekdayNames.js
@@ -1,0 +1,8 @@
+export const weekdayNames = {
+  'en-GB': {
+    Sunday: {
+      long: ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'],
+      short: ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'],
+    },
+  },
+};
