+++ packages/input-datepicker/test-helpers.js
@@ -1,0 +1,1 @@
+export { DatepickerInputObject } from './test-helpers/DatepickerInputObject.js';
