+++ packages/localize/test-helpers.js
@@ -1,9 +1,7 @@
-import { localize } from './src/localize.js';
-
-export const localizeTearDown = () => {
-  // makes sure that between tests the localization is reset to default state
-  localize._teardownHtmlLangAttributeObserver();
-  document.documentElement.lang = 'en-GB';
-  localize._setupHtmlLangAttributeObserver();
-  localize.reset();
-};
+export {
+  fakeImport,
+  resetFakeImport,
+  setupEmptyFakeImportsFor,
+  setupFakeImport,
+} from './test-helpers/fake-imports.js';
+export { localizeTearDown } from './test-helpers/localizeTearDown.js';
