+++ packages/validate/test-helpers.js
@@ -1,0 +1,1 @@
+export { smokeTestValidator } from './test-helpers/smokeTestValidator.js';
