+++ packages/validate/test-helpers/smokeTestValidator.js
@@ -1,8 +1,9 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { expect } from '@open-wc/testing';
 
 export const smokeTestValidator = (name, validator, value, params = undefined) => {
   const generated = validator(params);
-  expect(generated[0](value, params)[name]).to.be.true;
+  expect(generated[0](value, params)[name]).to.equal(true);
   if (params) {
     expect(generated[1]).to.equals(params);
   }
